create build dir if it does not exist

+++ site-builder/builder.ts
@@ -253,6 +253,7 @@ function main() {
     hljs.registerLanguage('rust', rust)
     hljs.registerLanguage('go', go)
     hljs.registerLanguage('ts', ts)
+    console.log("0")
 
     const program = new Command()
     program
@@ -266,10 +267,20 @@ function main() {
         .option("-c, --clean", "remove all files in build dir before build", true)
         .action((root, buildDir, options) => {
             if (options.clean) {
-                for (const entry of fs.readdirSync(buildDir)) {
-                    fs.rmSync(path.join(buildDir, entry), { recursive: true, force: true })
+                if (fs.existsSync(buildDir)) {
+                    if (fs.statSync(buildDir).isDirectory()) {
+                        console.log("3")
+                        for (const entry of fs.readdirSync(buildDir)) {
+                            fs.rmSync(path.join(buildDir, entry), { recursive: true, force: true })
+                        }
+                    } else {
+                        throw `not a directory: ${buildDir}`
+                    }
+                } else {
+                    fs.mkdirSync(buildDir)
                 }
             }
+            console.log("5")
             build(root, buildDir)
         })
